+++ 3D-Engine/Maths/Extensions.cs
@@ -4,12 +4,12 @@ namespace _3D_Engine
 {
     public static class Extensions
     {
-        // Double extensions
-        internal static byte Round_to_Byte(this double num) => (byte)((num >= 0) ? (num + 0.5) : (num - 0.5));
-        internal static int Round_to_Int(this double num) => (int)((num >= 0) ? (num + 0.5) : (num - 0.5));
+        // Float extensions
+        internal static byte Round_to_Byte(this float num) => (byte)((num >= 0) ? (num + 0.5) : (num - 0.5));
+        internal static int Round_to_Int(this float num) => (int)((num >= 0) ? (num + 0.5) : (num - 0.5));
 
         // Color extensions
-        public static Color Brighten(this Color colour, double fraction)
+        public static Color Brighten(this Color colour, float fraction)
         {
             fraction++;
 
@@ -26,16 +26,16 @@ namespace _3D_Engine
             return Color.FromArgb(new_a, new_r, new_g, new_b);
         }
 
-        public static Color Brighten_Percentage(this Color colour, double percentage) => Brighten(colour, percentage / 100);
-        public static Color Darken(this Color colour, double fraction) => Brighten(colour, fraction - 1);
-        public static Color Darken_Percentage(this Color colour, double percentage) => Darken(colour, percentage / 100);
+        public static Color Brighten_Percentage(this Color colour, float percentage) => Brighten(colour, percentage / 100);
+        public static Color Darken(this Color colour, float fraction) => Brighten(colour, fraction - 1);
+        public static Color Darken_Percentage(this Color colour, float percentage) => Darken(colour, percentage / 100);
 
         public static Color Mix(this Color colour, Color mixing_colour)
         {
-            byte new_a = Round_to_Byte(0.5 * (colour.A + mixing_colour.A));
-            byte new_r = Round_to_Byte(0.5 * (colour.R + mixing_colour.R));
-            byte new_g = Round_to_Byte(0.5 * (colour.G + mixing_colour.G));
-            byte new_b = Round_to_Byte(0.5 * (colour.B + mixing_colour.B));
+            byte new_a = Round_to_Byte(0.5f * (colour.A + mixing_colour.A));
+            byte new_r = Round_to_Byte(0.5f * (colour.R + mixing_colour.R));
+            byte new_g = Round_to_Byte(0.5f * (colour.G + mixing_colour.G));
+            byte new_b = Round_to_Byte(0.5f * (colour.B + mixing_colour.B));
 
             return Color.FromArgb(new_a, new_r, new_g, new_b);
         }
