+++ 3D-Engine/Scene/Scene Objects/Meshes/2D/Line.cs
@@ -27,7 +27,7 @@
             }
         }
 
-        public double Length { get; set; } //ss
+        public float Length { get; set; } //ss
         public Vector3D Unit_Vector { get; set; } //s
 
         #endregion
@@ -52,7 +52,7 @@
             Draw_Faces = false;
         }
 
-        public Line(Vector3D start_position, Vector3D unit_vector, double length) : this(start_position, start_position + unit_vector * length)
+        public Line(Vector3D start_position, Vector3D unit_vector, float length) : this(start_position, start_position + unit_vector * length)
         {
             Length = length;
             Unit_Vector = unit_vector;
