+++ 3D-Engine/Scene/Scene Objects/Meshes/3D/Sphere.cs
@@ -6,7 +6,7 @@ namespace _3D_Engine
     {
         #region Fields and Properties
 
-        public double Radius { get; set; }
+        public float Radius { get; set; }
         public int Res_Lat { get; set; }
         public int Res_Long { get; set; }
 
@@ -14,15 +14,15 @@ namespace _3D_Engine
 
         #region Constructors /////////////////////////////////////////
 
-        public Sphere(Vector3D origin, Vector3D direction_forward, Vector3D direction_up, double radius, int res_lat, int res_long) : base(origin, direction_forward, direction_up)
+        public Sphere(Vector3D origin, Vector3D direction_forward, Vector3D direction_up, float radius, int res_lat, int res_long) : base(origin, direction_forward, direction_up)
         {
             Dimension = 3;
 
             Radius = radius;
 
-            double angle_lat = 2 * Math.PI / res_lat, angle_long = Math.PI / (2 * res_long); // divide by 2?
+            float angle_lat = (float)2 * (float)Math.PI / res_lat, angle_long = (float)Math.PI / (2 * res_long); // divide by 2?
 
-            double x, y, z;
+            float x, y, z;
 
             Vertices = new Vertex[res_lat * res_long];
 
@@ -30,9 +30,9 @@ namespace _3D_Engine
             {
                 for (int j = 0; j < res_long; j++)
                 {
-                    x = radius * Math.Cos(angle_long * j) * Math.Sin(angle_lat * i);
-                    y = radius * Math.Sin(angle_long * j);
-                    z = radius * Math.Cos(angle_long * j) * Math.Cos(angle_lat * i);
+                    x = radius * (float)Math.Cos(angle_long * j) * (float)Math.Sin(angle_lat * i);
+                    y = radius * (float)Math.Sin(angle_long * j);
+                    z = radius * (float)Math.Cos(angle_long * j) * (float)Math.Cos(angle_lat * i);
                     Vertices[i * res_lat + j] = new Vertex(new Vector4D(x, y, z));
                 }
             }
