+++ 3D-Engine/Scene/Scene Objects/Meshes/3D/Text3D.cs
@@ -5,15 +5,15 @@
         #region Fields and Properties
 
         public string[] Fonts { get; set; }
-        public double Size { get; set; }
+        public float Size { get; set; }
         public char Style { get; set; }
-        public double Depth { get; set; }
+        public float Depth { get; set; }
 
         #endregion
 
         #region Constructors
 
-        public Text3D(Vector3D origin, Vector3D direction_forward, Vector3D direction_up, string[] fonts, double size, char style, double depth) : base(origin, direction_forward, direction_up)
+        public Text3D(Vector3D origin, Vector3D direction_forward, Vector3D direction_up, string[] fonts, float size, char style, float depth) : base(origin, direction_forward, direction_up)
         {
             Dimension = 3;
